chore: rename particle renderer
