+++ packages/core/src/particle/index.ts
@@ -1,1 +1,1 @@
-export { ParticleRenderer } from "./particle";
+export { ParticleRenderer } from "./particleRenderer";
